chore(compass-e2e-tests): Move the cheese around
