+++ packages/compass-e2e-tests/package.json
@@ -5,7 +5,7 @@
   "description": "E2E test suite for Compass app that follows smoke tests / feature testing matrix",
   "scripts": {
     "pretest": "mongodb-runner start --port=27018 --dbpath ./.mongodb",
-    "test": "xvfb-maybe mocha suites/**/*.test.js",
+    "test": "xvfb-maybe --auto-servernum -- node index.js",
     "posttest": "mongodb-runner stop --port=27018 && node -e \"fs.rmdirSync('.mongodb', { recursive: true })\"",
     "test-ci": "npm run test",
     "posttest-ci": "node ../../scripts/killall-mongo.js",
@@ -19,6 +19,7 @@
     "@types/webdriverio": "^4.13.3",
     "chai": "^4.3.4",
     "electron-rebuild": "^2.3.5",
+    "glob": "^7.1.6",
     "mocha": "^8.4.0",
     "mongodb-runner": "^4.8.3",
     "spectron": "^8.0.0",
